Fix 32 bits build
Dividing a u64 in 32bits architecture would produce the error :
	vd55g1.c:(.text+0x1e38): undefined reference to `__udivdi3'
Change mipi_freq variable type to u32 instead of using div helpers as it
is enough to store its value anyway.

+++ vd55g1.c
@@ -949,7 +949,7 @@ static int vd55g1_prepare_clock_tree(struct vd55g1 *sensor)
 {
 	struct i2c_client *client = sensor->i2c_client;
 	/* Double data rate */
-	u64 mipi_freq = link_freq[0] * 2;
+	u32 mipi_freq = link_freq[0] * 2;
 	u32 sys_clk, mipi_div, pixel_div;
 	int ret = 0;
 
@@ -964,7 +964,7 @@ static int vd55g1_prepare_clock_tree(struct vd55g1 *sensor)
 	if (mipi_freq < 250 * HZ_PER_MHZ ||
 	    mipi_freq > 1200 * HZ_PER_MHZ) {
 		dev_err(&client->dev,
-			"Only 250Mhz-1200Mhz link frequency range supported. Provided %llu MHz\n",
+			"Only 250Mhz-1200Mhz link frequency range supported. Provided %lu MHz\n",
 			mipi_freq / HZ_PER_MHZ);
 		return -EINVAL;
 	}
